Make generating random NeuralNetwork for Entities
Now every entity has a NeuralNetwork with default quantity of layers(3) and neurons in each of them(also 3). Input layer, hidden layer and output layer. When new entity is borned it has a link to a mother entity NeuralNetwork, need to fix in future with making mutating. Now every Neuron has type.